backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
qdrant-client==1.7.0
sentence-transformers==2.2.2
huggingface-hub==0.19.4
gitpython==3.1.40
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.2
typing-extensions==4.8.0
